==> s1/1_first of all.ipynb <==
##First Of All

商品子集都是偏服务类的商品，涵盖阿里巴巴集团十个主要的商品大类，例如汽车售后服务、摄影服务、餐饮、电影等，其特色是线上购买、线下服务。

指导性建议

1. 把预测问题看成一个二分类问题：第一类，购买；第二类，不购买
2. 确定样本的粒度：(user_id,item_id)；
> 如(6824287,25076172), (6824287,3507617)就是不同的样本

3. 构建训练集、测试集：训练集和测试集的时间点是一个关键因素，选择确定之后整个竞赛过程就不要反复调整了；请谨慎选择他们和target天的关系，这是很有讲究的，。一个最基本知识就是不要把训练集和测试集选到同一天。
> 3.1 构建验证集：在你做到一定水平之后，你就会发现只有训练集和测试集是不够的，你可能会想到选择第三个日期来构建验证集。构建验证集有这样几个好处：防止测试集被分析的太多了反而忘了线上集的分数才是你最终的排名标准；利用测试集和验证集的时间关系更好的猜到线上集上能拿到的成绩;把验证集、训练集、测试集用各种方式换过来试一下，也许你能看到你原来想象不到的现象出现

4. 生成特征：一个最基础的特征组合包括UI特征，U特征，I特征；特征的具体定义就是发挥你智慧的空间了
5. 选择一个模型：别让模型训练的太慢了，这会影响你的工作效率
6. 构建提交候选集（线上集）：对于在target天可能产生购买行为的（U，I），我们虽然无法知道他到底最终是否被购买了，但是我们要为它们产生特征并交由模型进行预测。？不太明白这句
> 6.1 不要试图预测从来没有产生行为的（U，I）：虽然没有产生过行为的（U，I）在最终结果集中占比很大，是一个非常诱人的部分；但是为了预测这部分样本，产生候选集的大小是以乘法级别上涨的。我们尝试过计算商品的相关，用户和商品的隐变量，最活跃的用户，最热销的商品这几种方法，都没有能够有效提升最终的F值。如果你没有崭新的思路或者更细的耕耘，请不要在这条路上浪费时间。

将**tianchi_mobile_recommend_train_user.cvs**划分为训练集和验证集，用于测试算法的效果。
> 如何划分？ 11-18～12-17作为训练集train_set，12-18作为验证集validate_set

推荐集 predict_set 提交的结果，how？
```
user_id,item_id
158978240,25076259
6824287,25076172
...
25076172,386660
```

##评估

$准确率 percision= \frac{hits}{pnums (我们推荐的条数)}$  
$召回率 recall=\frac{hits}{vnums (线上验证集的条数)}$  
$ F1 = \frac{2*P*R}{(P+R)} = \frac{2*hits}{(pnums + vnums)}$ 最终的结果

所以要获得较好的F1需要平衡推荐条数和准确率之间的关系

##tips

* 对数据的处理都写成函数的形式，方便自动化
