==> s1/2_dt.ipynb <==
##训练集(train_set) / 验证集(validate_set) / 推荐集(predict_set)

3. 构建训练集、测试集：训练集和测试集的时间点是一个关键因素，选择确定之后整个竞赛过程就不要反复调整了；请谨慎选择他们和target天的关系，这是很有讲究的，。一个最基本知识就是不要把训练集和测试集选到同一天。
> 1. 训练集和测试集在形式上的区别在于前者有类别标签，后者则需要模型输出相应的类别标签。
> 2. 构建验证集：在你做到一定水平之后，你就会发现只有训练集和测试集是不够的，你可能会想到选择第三个日期来构建验证集。构建验证集有这样几个好处：防止测试集被分析的太多了反而忘了线上集的分数才是你最终的排名标准；利用测试集和验证集的时间关系更好的猜到线上集上能拿到的成绩;把验证集、训练集、测试集用各种方式换过来试一下，也许你能看到你原来想象不到的现象出现

6. 构建提交候选集（线上集）：对于在target天可能产生购买行为的（U，I），我们虽然无法知道他到底最终是否被购买了，但是我们要为它们产生特征并交由模型进行预测。？不太明白这句
> 6.1 不要试图预测从来没有产生行为的（U，I）：虽然没有产生过行为的（U，I）在最终结果集中占比很大，是一个非常诱人的部分；但是为了预测这部分样本，产生候选集的大小是以乘法级别上涨的。我们尝试过计算商品的相关，用户和商品的隐变量，最活跃的用户，最热销的商品这几种方法，都没有能够有效提升最终的F值。如果你没有崭新的思路或者更细的耕耘，请不要在这条路上浪费时间。

将**tianchi_mobile_recommend_train_user.cvs**划分为训练集和验证集，用于测试算法的效果。
> 如何划分？   
> 样本的粒度：(user_id,item_id)  
> 训练集train_set，11-18～12-17，其相应的类别标签（tag：1,购买; 0,未购买）来自于12-18日的购买情况，各个特征以此时间段计算出  
> 测试集test_set，11-18～12-18，没有类别标签，特征由此段时间计算出  
> 验证集validate_set，12-18，这一天的购买了的数据，字段只要包含(user_id,item_id)  
> 候选集predict_set如何产生


import pandas as pd
import numpy as np

df_train_user = pd.read_csv('data/raw/tianchi_mobile_recommend_train_user.csv')
df_train_user

# # 过滤掉 12.12 的记录
# df_train_user = df_train_user[df_train_user.time.map(lambda x: not x.startswith('2014-12-12'))]
# df_train_user.to_csv('data/raw/no1212tianchi_mobile_recommend_train_user.csv', index=False)

分割点为 2014-12-18 00  
训练集train_set，11-18～12-17，其相应的类别标签（tag：1,购买; 0,未购买）来自于12-18日的购买情况，各个特征以此时间段计算出  
测试集test_set，11-18～12-18，没有类别标签，特征由此段时间计算出  
验证集validate_set，12-18，这一天的购买了的数据，字段只要包含(user_id,item_id) 候选集predict_set如何产生  

import datetime

def ge_validate_set(df, separate_day='2014-12-18 00', to_file='data/raw_validate_set.csv'):
    # TODO: 使用datetime格式替代str格式来比较
    # separate_day = datetime.datetime(2014,12,18,0)
    # parseDate = lambda s: datetime.datetime.strptime(s, "%Y-%m-%d %H")
    validate_set = df[ df.time >= '2014-12-18 00']
    validate_set = validate_set[ df.behavior_type > 3]
    # TODO: 去重
    if to_file:
        validate_set.to_csv(to_file, index=False)
    return validate_set


def ge_result(df=None, to_file='data/result.csv'):
    # 结果只保留 user_id, item_id
    result = df or pd.read_csv('data/raw_validate_set.csv')
    #result = result.drop(['behavior_type','user_geohash','item_category','time'], axis=1)
    result = result[['user_id', 'item_id']]
    # 过滤出在P商品子集上的记录
    P = pd.read_csv('data/raw/tianchi_mobile_recommend_train_item.csv')
    P_item_set = set(P.item_id.values)
    result = result[result.item_id.map(lambda x: x in P_item_set)]
    # 去重
    result = result.groupby(['user_id', 'item_id'])
    result = pd.DataFrame(result.groups.keys(), columns=['user_id', 'item_id'])
    if to_file:
        result.to_csv(to_file, index=False)
    return result


# train_set：
# 1. 过滤出time<2014-12-18 00的所有记录
# 2. 对过滤出的记录添加buy column(根据12-18号的记录对其进行标记)
# 3. 要把time字段替换为date_delta -> 放到特征工程中再做吧
def ge_train_set(df, result, separate_day='2014-12-18 00', to_file='data/1212raw_train_set.csv'):
    # TODO: 使用datetime格式替代str格式来比较
    # separate_day = datetime.datetime(2014,12,18,0)
    # parseDate = lambda s: datetime.datetime.strptime(s, "%Y-%m-%d %H")
    result_set = set((i[0],i[1]) for i in result.values)
    train_set = df[ df.time < '2014-12-18 00']
    
    if to_file:
        train_set.to_csv(to_file, index=False)
    return train_set


def ge_test_set(df, to_file='data/test_set/tianchi_mobile_recommend_test_user.csv'):
    # 商品子集 P
    P = pd.read_csv('data/raw/tianchi_mobile_recommend_train_item.csv')
    P_item_set = set(P.item_id.values)
    # 过滤出 item_id 属于 P(商品子集) 的记录
    test_set = df[df.item_id.map(lambda x: x in P_item_set)]
    if to_file:
        test_set.to_csv(to_file, index=False)
    return test_set

###验证集validate_set

12-18，这一天的购买了的数据，字段只要包含(user_id,item_id) 候选集predict_set如何产生

validate_set = ge_validate_set(df_train_user)
validate_set

#result = ge_result(validate_set)
result = ge_result()
result

###训练集train_set

11-18～12-17，其相应的类别标签（tag：1,购买; 0,未购买）来自于12-18日的购买情况，各个特征以此时间段计算出

train_set = ge_train_set(df_train_user, result)
train_set

###测试集test_set

test_set = ge_test_set(df_train_user)
test_set
