==> s1/2.1_undersample.ipynb <==
##随机采样

from sklearn.random_projection import GaussianRandomProjection
from sklearn import datasets
import numpy as np
from UnbalancedDataset import UnbalancedDataset
import pandas as pd

#r = UnbalancedDataset.UnderSampler()

#r.fit([[7],[8],[9],[0]], [0,0,0,1])

#r.resample()

def ge_undersample_train_set(f='data/train_set/train_set.csv', ratio=2000, 
                          to_file='data/train_set/train_set_undersample_%d.csv'):
    train_set = pd.read_csv(f)
    gp = train_set.groupby(['buy'])['buy']
    positive_sample_idxs = np.array(gp.groups[1])
    negative_sample_idxs = gp.groups[0]
    # 对负样本进行随机下采样，正负样本比 1:ratio
    ratio = ratio
    negative_sample_sampling_idxs = np.random.choice(negative_sample_idxs, ratio*len(positive_sample_idxs), replace=False)
    idxs = np.concatenate((negative_sample_sampling_idxs, positive_sample_idxs))
    # 对正负样本idx进行洗牌
    np.random.shuffle(idxs)
    undersample_train_set = train_set.ix[idxs]
    if to_file:
        undersample_train_set.to_csv(to_file%ratio)
    return undersample_train_set

# 1212
ge_undersample_train_set('data/train_set/1212train_set.csv', ratio=14,
                      to_file='data/train_set/1212train_set_undersample_%d.csv')
# ~1212
ge_undersample_train_set(ratio=1200)

#train_set = pd.read_csv('data/train_set/train_set.csv')
#train_set[train_set.buy > 0]

#datasets.samples_generator

#np.random.uniform(1,9, 8)
