==> s1/3.1_behavior_count.ipynb <==
import pandas as pd

from collections import defaultdict
import datetime

_features = {
    'f_behavior_count': {1:'f3_1', 2:'f3_2', 3:'f3_3', 4:'f3_4'},
    'f_behavior_lasttime_delta': {1:'f3_5', 2:'f3_6', 3:'f3_7', 4:'f3_8'},
}

def f_behavior_count(df):
    '''截止到最后一天，用户对商品进行过访问的次数、购买的次数、收藏的次数、购物车的次数
    TODO: 加时间衰减
    @return dict; {'f3_1':browse_count, 'f3_2': collect_count, ...}
    '''
    #features = {1:'f3_1', 2:'f3_2', 3:'f3_3', 4:'f3_4'}
    #f = defaultdict(int)
    f = [0,0,0,0]
    c = df.groupby('behavior_type')['behavior_type'].count()
    for idx in c.index:
        f[idx-1] = c[idx]
    #return [f['f3_1'],f['f3_2'], f['f3_3'], f['f3_4']]
    return f

##训练集features

raw_train_set = pd.read_csv('data/1212raw_train_set.csv')
raw_train_set = raw_train_set[['user_id', 'item_id', 'behavior_type']]
gp_ui = raw_train_set.groupby(['user_id', 'item_id'])
print 'group by user_id, item_id OK'

def l(gp_ui) :
    train_list = []
    c = 0
    for (u,i), gpdf in gp_ui:
        #train_set.loc[c] = f_behavior_count(gpdf)
        train_list.append(f_behavior_count(gpdf))
        c += 1
        if c%10000==0: print c
    print 'done'
    return train_list

train_list = l(gp_ui)
columns = []#['user_id', 'item_id']
columns.extend(_features['f_behavior_count'].values())
train_set = pd.DataFrame(train_list, columns=columns)
train_set.to_csv('data/train_set/1212feature_set_behavior_count.csv', index=False)

##测试集features

raw_test_set = pd.read_csv('data/test_set/tianchi_mobile_recommend_test_user.csv')
raw_test_set = raw_test_set[['user_id', 'item_id', 'behavior_type']]

gp_ui = raw_test_set.groupby(['user_id', 'item_id'])
print 'group by user_id, item_id OK'

def l(gp_ui) :
    test_list = []
    c = 0
    for (u,i), gpdf in gp_ui:
        #train_set.loc[c] = f_behavior_count(gpdf)
        test_list.append(f_behavior_count(gpdf))
        c += 1
        if c%10000==0: print c
    print 'done'
    return test_list

test_list = l(gp_ui)
columns = []#['user_id', 'item_id']
columns.extend(_features['f_behavior_count'].values())
test_set = pd.DataFrame(test_list, columns=columns)
test_set.to_csv('data/test_set/1212feature_set_behavior_count.csv', index=False)
