==> s1/3.2_behavior_lasttime_delta.ipynb <==
import pandas as pd

from collections import defaultdict
import datetime

_features = {
    'f_behavior_count': {1:'f3_1', 2:'f3_2', 3:'f3_3', 4:'f3_4'},
    'f_behavior_lasttime_delta': {1:'f3_5', 2:'f3_6', 3:'f3_7', 4:'f3_8'},
}

parseDate = lambda s: datetime.datetime.strptime(s, "%Y-%m-%d %H")
def f_behavior_lasttime_delta(df, seperator=datetime.datetime(2014,12,19,0)):
    '''用户最后一次对商品访问（收藏、购物车、购买）距离分隔日期（如62）的时间间隔，如果没有发生动作，默认间隔60天
    @return dict; 
    '''
#     features = {1:'f3_5', 2:'f3_6', 3:'f3_7', 4:'f3_8'}
#     f = defaultdict(int)
#     for v in features.values(): f[v] = 60
    f = [60, 60, 60, 60]
    group_max = df.groupby('behavior_type')['time'].max()
    for idx in group_max.index:
        f[idx-1] = (seperator - datetime.datetime.strptime(group_max[idx], "%Y-%m-%d %H")).days
#     return [f['f3_5'],f['f3_6'], f['f3_7'], f['f3_8']]
    return f

##训练集features

raw_train_set = pd.read_csv('data/1212raw_train_set.csv')
raw_train_set = raw_train_set[['user_id', 'item_id', 'behavior_type', 'time']]
gp_ui = raw_train_set.groupby(['user_id', 'item_id'])
print 'group by user_id, item_id OK'

def l(gp_ui) :
    train_list = []
    c = 0
    for (u,i), gpdf in gp_ui:
        #train_set.loc[c] = f_behavior_count(gpdf)
        train_list.append(f_behavior_lasttime_delta(gpdf))
        c += 1
        if c%10000==0: print c
    print 'done'
    return train_list

train_list = l(gp_ui)
columns = []#['user_id', 'item_id']
columns.extend(_features['f_behavior_lasttime_delta'].values())
train_set = pd.DataFrame(train_list, columns=columns)
train_set.to_csv('data/train_set/1212feature_set_behavior_lasttime_delta.csv', index=False)

##测试集features

# raw_test_set = pd.read_csv('data/test_set/tianchi_mobile_recommend_test_user.csv')
# raw_test_set = raw_test_set[['user_id', 'item_id', 'behavior_type', 'time']]
# gp_ui = raw_test_set.groupby(['user_id', 'item_id'])
# print 'group by user_id, item_id OK'

# # 测试集的分割点为 12-19 00
# seperator=datetime.datetime(2014,12,20,0)

# def l(gp_ui) :
#     test_list = []
#     c = 0
#     for (u,i), gpdf in gp_ui:
#         #train_set.loc[c] = f_behavior_count(gpdf)
#         test_list.append(f_behavior_lasttime_delta(gpdf, seperator))
#         c += 1
#         if c%10000==0: print c
#     print 'done'
#     return test_list

# test_list = l(gp_ui)
# columns = []#['user_id', 'item_id']
# columns.extend(_features['f_behavior_lasttime_delta'].values())
# test_set = pd.DataFrame(test_list, columns=columns)
# test_set.to_csv('data/test_set/1212feature_set_behavior_lasttime_delta.csv', index=False)
