==> s1/3.3_user_buyCount.ipynb <==
import numpy as np
import pandas as pd
import patsy
import cPickle

##2 比值类特征

y用户购买该item次数/总购买次数

def user_buy_count(df='data/raw_train_set.csv', ui='data/train_set/ui_set.csv', 
                   to_file='data/train_set/feature_u_count.csv'):
    '''U特征，基本统计量特征
    购买量 f2_1, 浏览量 f2_2
    '''
#     df='data/raw_train_set.csv'
#     ui='data/train_set/ui_set.csv'
#     to_file='data/train_set/buy_count.csv'
    df = pd.read_csv(df)
    ui = pd.read_csv(ui)
    # 购买量 f2_1
    buyset = df[df.behavior_type == 4]
    c = buyset.groupby(['user_id'])['user_id'].count()
    buy_count = c.to_dict()
    buy_count_list = np.array(ui.user_id.map(lambda x: buy_count.get(x, 0)))
    buy_count_list.shape = -1, 1
    
    count_set = pd.DataFrame(buy_count_list, columns=['f2_1'])
    
    # 浏览量 f2_2
    browse_set = df[df.behavior_type == 1]
    c = browse_set.groupby(['user_id'])['user_id'].count()
    browse_count = c.to_dict()
    count_set['f2_2'] = np.array(ui.user_id.map(lambda x: browse_count.get(x, 0)))
    if to_file:
        count_set.to_csv(to_file, index=False)
    return count_set
    

##训练集 train_set

# 1212
user_buy_count('data/1212raw_train_set.csv', 'data/train_set/1212ui_set.csv', 
                           'data/train_set/1212feature_u_count.csv')
# ～1212
# user_buy_count(to_file='data/train_set/feature_u_count.csv')

#count_set[count_set.f2_2 == 2]

##测试集 test_set

# 1212
user_buy_count('data/raw/tianchi_mobile_recommend_train_user.csv', 
               'data/test_set/1212ui_set.csv', 
               'data/test_set/1212feature_u_count.csv')
# # ~1212
# user_buy_count('data/raw/no1212tianchi_mobile_recommend_train_user.csv', 
#                'data/test_set/ui_set.csv', 
#                'data/test_set/feature_u_count.csv')
