==> s1/3.4_item_saleCount.ipynb <==
import numpy as np
import pandas as pd
import datetime

## chat

嵌入ipython notebook的chat room, websocket实现，注册，创建房间号

def item_sale_count(df='data/1212raw_train_set.csv', 
                   seperator=datetime.datetime(2014,12,18,0),
                   ui='data/train_set/1212ui_set.csv',
                   to_file='data/train_set/1212feature_i_count.csv'):
    '''I特征，基本统计量特征
    销售量 f1_1_3, f1_1_7 周期为3, 7
    购买人数（周期内经过去重的购买人数）f1_2_3, f1_2_7
    某一周期内多次购买的用户数 f1_3_3 f1_3_7
    返客率（某一周期内多次购买的用户数/总购买用户数）f1_3_3 / f1_2_3
    '''
#     df='data/raw_train_set.csv'
#     ui='data/train_set/ui_set.csv'
#     to_file='data/train_set/buy_count.csv'

    # 分隔日期前3、7天
    p3 = datetime.datetime.strftime(seperator - datetime.timedelta(3), '%Y-%m-%d %H')
    p7 = datetime.datetime.strftime(seperator - datetime.timedelta(7), '%Y-%m-%d %H')
    sale_p3_count = {}
    sale_p7_count = {}
    num_of_people_p3_count = {}
    num_of_people_p7_count = {}
    repeat_buy_user_p3_count = {}
    repeat_buy_user_p7_count = {}
    # TODO
    #_features = {
    # 'f1_1_3': sale_p3_count,
    #}
    
    df = pd.read_csv(df)
    ui = pd.read_csv(ui)
    # 销售量 f1_1
    saleset = df[df.behavior_type == 4]
    gp = saleset.groupby(['item_id'])[['item_id', 'time']] # 大概 90,000个
    c = 0
    for i, gpdf in gp:
        is_p3 = gpdf.time >= p3 #pandas.core.series.Series
        is_p7 = gpdf.time >= p7
        user_in_p3 = gpdf[is_p3]['user_id']
        user_in_p7 = gpdf[is_p7]['user_id']
        
        # 销售量 f1_1_3, f1_1_7 周期为3, 7
        sale_p3_count[i] = is_p3.values.sum()
        sale_p7_count[i] = is_p7.values.sum()
        # 购买人数（周期内经过去重的购买人数）f1_2_3, f1_2_7
        num_of_people_p3_count[i] = user_in_p3.unique().shape[0]
        num_of_people_p7_count[i] = user_in_p7.unique().shape[0]
        #某一周期内多次购买的用户数 f1_3_3 f1_3_7
        repeat_buy_user_p3_count[i] = user_in_p3[user_in_p3.duplicated()].unique().shape[0]
        repeat_buy_user_p7_count[i] = user_in_p7[user_in_p7.duplicated()].unique().shape[0]
        
        c += 1
        if c%10000 == 0: print c

    add_i_feature = lambda count_set: \
                        ui.item_id.map(lambda x: count_set.get(x, 0)) # 有种写lisp的感觉
    # 记录到 sale_set
    sale_set = pd.DataFrame()
    sale_set['f1_1_3'] = add_i_feature(sale_p3_count)
    sale_set['f1_1_7'] = add_i_feature(sale_p7_count)
    sale_set['f1_2_3'] = add_i_feature(num_of_people_p3_count)
    sale_set['f1_2_7'] = add_i_feature(num_of_people_p7_count)
    sale_set['f1_3_3'] = add_i_feature(repeat_buy_user_p3_count)
    sale_set['f1_3_7'] = add_i_feature(repeat_buy_user_p7_count)
    
    if to_file:
        sale_set.to_csv(to_file, index=False)
    return sale_set

##训练集 train_set

item_sale_count()

##测试集 test_set

item_sale_count(
    df='data/test_set/tianchi_mobile_recommend_test_user.csv', 
    seperator=datetime.datetime(2014,12,19,0),
    ui='data/test_set/1212ui_set.csv',
    to_file='data/test_set/1212feature_i_count.csv')
