==> s1/4.0_models.ipynb <==
import numpy as np
import pandas as pd
from sklearn import cross_validation
from sklearn.ensemble import GradientBoostingClassifier
from sklearn.ensemble import GradientBoostingRegressor
from sklearn.grid_search import GridSearchCV
import patsy

def diff_sets(s1, s2):
    s1 = set(map(tuple, s1.values.tolist()))
    s2 = set(map(tuple, s2.values.tolist()))
    return s1 & s2


def evaluate(p, v):
    #p = pd.read_csv('data/output/lr/tianchi_mobile_recommendation_predict.csv')
    #v = pd.read_csv('data/result.csv')
    
    pset = set(map(tuple, p.values.tolist()))
    vset = set(map(tuple, v.values.tolist()))
    hits_set = pset & vset
    
    hits  = float(len(hits_set))
    pnums = float(len(pset))
    vnums = float(len(vset))
    
    print 'F1：', format(2 * hits / (pnums + vnums), '.2%'),
    print '\t准确率：', format((hits / pnums), '.2%'),
    print '\t召回率：', format(hits / vnums, '.2%'),
    print '\thits：%d'%hits

def get_recommend_set(df, pnums=461):
    return df.sort(columns=['buy'], ascending=False)[:pnums][['user_id', 'item_id']]

http://scikit-learn.org/stable/modules/cross_validation.html

#cross_validation.cross_val_score??

est = GradientBoostingClassifier(n_estimators=400, max_depth=4, min_samples_leaf=2)

# #est.score(X_test, y_test)
# n_samples = X.shape[0]
# cv = cross_validation.ShuffleSplit(n_samples, n_iter=1, test_size=0.25)
# scores = cross_validation.cross_val_score(est, X, y, cv=cv, scoring='f1_weighted', n_jobs=1)

est.fit(X_train, y_train[:,2])

output_prob = est.predict_proba(X_test)
df_y_test = pd.DataFrame(y_test, columns=['user_id','item_id', 'buy'])
y_predict = df_y_test[['user_id', 'item_id']]
y_predict['buy'] = output_prob[0:, 1]

result = df_y_test[df_y_test.buy > 0][['user_id','item_id']]
print 'result %d'%result.shape[0]
recommend_set = get_recommend_set(y_predict, 500)
evaluate(recommend_set, result)
