==> s1/4.1_LR.ipynb <==
##logistic回归

import pandas as pd
import numpy as np
import statsmodels.api as sm # statistical modeling and econometrics in Python
from statsmodels.nonparametric.kde import KDEUnivariate
from statsmodels.nonparametric import smoothers_lowess
import patsy
import time

train_set = pd.read_csv('data/train_set/1212train_set.csv')

# 模型公式
# 写成用来预测Survived的公式
# The C() lets our regression know that those variables are categorical.
formula = 'buy ~ f3_1 + f3_2 + f3_3 + f3_4 + \
                 f3_5 + f3_6 + f3_7 + f3_8 + \
                 f1_1_3 + f1_1_7 + f1_2_3 + f1_2_7 + f1_3_3 + f1_3_7 + \
                 f2_1 + np.true_divide(f3_4, f2_1+0.01)'# + np.true_divide(f2_1, f2_2)'
train_set['intercept'] = 1.0
# 用patsy的dmatrices生成一个对regression友好的dataframe
y, x = patsy.dmatrices(formula, data=train_set, return_type='dataframe')

# 实例化一个logistic regression模型
model = sm.Logit(y, x)

# fit our model to the training data
# res.predict(x) 用于预测
res = model.fit()

res.summary()

# 保存结果，用于之后的预测
results = {'Logit':[res, formula]}
# 结果保存到文件
res.save('models/lr [%s].pickle'%formula, remove_data=True)
del train_set

###预测

# now = time.strftime('%Y-%m-%d %H:%M',time.localtime())
# # 用patsy的dmatrices生成一个对regression友好的dataframe
# test_set = pd.read_csv("data/test_set/1212test_set.csv")
# test_set['buy'] = 0
# y, x = patsy.dmatrices(formula, data = test_set, return_type='dataframe')

# predict_set = pd.read_csv("data/test_set/1212ui_set.csv")

# print 'predicting ...'
# predict_set['buy'] = res.predict(x)
# print 'predict done'

# predict_set.to_csv("data/output/lr/predict_set.csv", index=False)

##过滤

# POINT = 0.2230 # 看summary定
# #filted_set = predict_set[ predict_set.buy > POINT ]
# #filted_set[['user_id','item_id']].to_csv("data/output/lr/tianchi_mobile_recommendation_predict.csv", index=False)

# # # 过滤出在P商品子集上的记录
# # P = pd.read_csv('data/raw/tianchi_mobile_recommend_train_item.csv')
# # P_item_set = set(P.item_id.values)
# # filted_set = predict_set[predict_set.item_id.map(lambda x: x in P_item_set)]

# # vnums是461
# recomm_set = predict_set.sort(columns=['buy'], ascending=False)[:461][['user_id','item_id']]
# recomm_set.to_csv("data/output/lr/tianchi_mobile_recommendation_predict.csv", index=False)
