==> s1/4.2_GBRT.ipynb <==
import pandas as pd
import numpy as np
from sklearn.ensemble import GradientBoostingClassifier
from sklearn.ensemble import GradientBoostingRegressor
from sklearn.grid_search import GridSearchCV
import cPickle
import patsy

import matplotlib.pyplot as plt
%matplotlib inline

Estimators support arguments to control the fitting behaviour -- these arguments are often called hyperparameters. Among the most important ones for GBRT are:

* number of regression trees (n_estimators)
* depth of each individual tree (max_depth)
* loss function (loss)
* learning rate (learning_rate)

For example if you want to fit a regression model with 100 trees of depth 3 using least-squares:

##Hyperparameter tuning

太耗时间了 12h+

# X = train_data[0:, 1:]
# y = train_data[0:, 0]

# param_grid = {
#     'learning_rate': [0.1, 0.05],
#     'max_depth': [4, 6],
#     'min_samples_leaf': [3, 9, 17],
#     #'max_features': [1.0, 0.3, 0.1]
# }
# est = GradientBoostingClassifier(n_estimators=500)
# gs_cv = GridSearchCV(est, param_grid).fit(X, y)
# # best hyperparameter setting
# gs_cv.best_params_

##训练数据

train_set = pd.read_csv('data/train_set/1212train_set.csv')

##训练

#train_data = train_set.values
formula = 'buy ~ f3_1 + f3_2 + f3_3 + f3_4 + \
                 f3_5 + f3_6 + f3_7 + f3_8 + \
                 f1_1_3 + f1_1_7 + f1_2_3 + f1_2_7 + f1_3_3 + f1_3_7 + \
                 f2_1 + np.true_divide(f3_4, f2_1+0.01)'# + np.true_divide(f2_1, f2_2)'
# 用patsy的dmatrices生成一个对 友好的dataframe
y, x = patsy.dmatrices(formula, data=train_set, return_type='dataframe')

x = x.values[0:, 1:]
y = y.values[0:, 0]

est = GradientBoostingClassifier(n_estimators=600, max_depth=6, min_samples_leaf=2)
est.fit(x, y)

f = open('models/f16_gbrt 1212[n_estimators=600,max_depth=6, min_samples_leaf=2].pickle', 'w')
cPickle.dump(est, f)
f.close()

##预测

test_set = pd.read_csv('data/test_set/1212test_set.csv')
#test_data = test_set.values
test_set['buy'] = 0

# formula = 'buy ~ f3_1 + f3_2 + f3_3 + f3_4 + \
#                  f3_5 + f3_6 + f3_7 + f3_8 + \
#                  f2_1 + np.true_divide(f3_4, f2_1+0.01)'# + np.true_divide(f2_1, f2_2)'
# 用patsy的dmatrices生成一个对 友好的dataframe
y, x = patsy.dmatrices(formula, data=test_set, return_type='dataframe')

x = x.values[:, 1:]
y = y.values[:, 0]

#output = est.predict(test_data[0:, 2:])
output_prob = est.predict_proba(x)

predict_set = pd.read_csv("data/test_set/1212ui_set.csv")

print 'predicting ...'
predict_set['buy'] = output_prob[0:,1]
print 'predict done'

predict_set.to_csv("data/output/gbrt/f16_predict_set 1212 600 6.csv", index=False)

##过滤

# predict_set = pd.read_csv('data/output/gbrt/predict_set.csv')

# recomm_set = predict_set.sort(columns=['buy'], ascending=False)[:421][['user_id','item_id']]
# recomm_set.to_csv("data/output/gbrt/tianchi_mobile_recommendation_predict.csv", index=False)

print 'ok'

##
