==> s1/5_F1.ipynb <==
##评估

$准确率 percision= \frac{hits}{pnums (我们推荐的条数)}$  
$召回率 recall=\frac{hits}{vnums (线上验证集的条数)}$  
$ F1 = \frac{2*P*R}{(P+R)} = \frac{2*hits}{(pnums + vnums)}$ 最终的结果

所以要获得较好的F1需要平衡推荐条数和准确率之间的关系

import pandas as pd
import matplotlib.pyplot as plt
%matplotlib inline

def diff_sets(s1, s2):
    s1 = set(map(tuple, s1.values.tolist()))
    s2 = set(map(tuple, s2.values.tolist()))
    return s1 & s2


def evaluate(p):
    #p = pd.read_csv('data/output/lr/tianchi_mobile_recommendation_predict.csv')
    v = pd.read_csv('data/result.csv')
    
    pset = set(map(tuple, p.values.tolist()))
    vset = set(map(tuple, v.values.tolist()))
    hits_set = pset & vset
    
    hits  = float(len(hits_set))
    pnums = float(len(pset))
    vnums = float(len(vset))
    
    print 'F1：', format(2 * hits / (pnums + vnums), '.2%'),
    print '\t准确率：', format((hits / pnums), '.2%'),
    print '\t召回率：', format(hits / vnums, '.2%'),
    print '\thits：%d'%hits

def get_recommed_set(df, pnums=461):
    return df.sort(columns=['buy'], ascending=False)[:pnums][['user_id', 'item_id']]

# predict_set = pd.read_csv('data/output/gbrt/predict_set.csv')

# # # 每个人 buy最大的ui
# # recomm_index = [gpdf.buy.idxmax() for u,gpdf in predict_set.groupby(['user_id'])]
# # recomm_set = predict_set.ix[recomm_index]
# # recomm_set = recomm_set.sort(columns=['buy'], ascending=False)[:421][['user_id','item_id']]
# # recomm_set.to_csv("data/output/lr/tianchi_mobile_recommendation_predict.csv", index=False)
# # evaluate(recomm_set)

# recomm_set = predict_set.sort(columns=['buy'], ascending=False)[:461][['user_id','item_id']]

# evaluate(recomm_set)

s1 = pd.read_csv('data/output/gbrt/f16_predict_set 1212 700 6.csv')
s2 = pd.read_csv('data/output/gbrt/f16_predict_set 1212 600 6.csv')
d2 = diff_sets(get_recommed_set(s1, 300),
               get_recommed_set(s2, 300))
v = pd.read_csv('data/result.csv')
len(d2)

# train_set = pd.read_csv('data/train_set/1212train_set.csv')
# train_set[train_set.buy > 0]

predict_set = pd.read_csv('data/output/gbrt/f16_predict_set 1212 600 6.csv')
recomm_set = predict_set.sort(columns=['buy'], ascending=False)[:300][['user_id','item_id']]
#recomm_set.to_csv("data/output/gbrt/tianchi_mobile_recommendation_predict-last100.csv", index=False)
evaluate(recomm_set)

p = predict_set.sort(columns=['buy'], ascending=False)[:300]
p
p[p.user_id == 54691351]

#evaluate(pd.read_csv('data/output/0.07375, 0.07375 (2015-04-15 22:26:14) gbrt400 1212.csv'))

# evaluate(pd.read_csv('data/test_set/test_set.csv')\
#          .sort(columns=['f3_7'], ascending=True)[:461][['user_id','item_id']])

##比较两个推荐集的差异

###1. 控制变量 有无1212

s400 = pd.read_csv('data/output/gbrt/predict_set~1212 500.csv')
s1212_400 = pd.read_csv('data/output/gbrt/predict_set 1212 500.csv')

d1 = diff_sets(get_recommed_set(s400), get_recommed_set(s1212_400))
len(d1)

###2. 控制变量 模型 （1212）

s1 = pd.read_csv('data/output/0.044, 0.4772 (2015-04-14 21:29:22) lr.csv')
s2 = pd.read_csv('data/output/0.07375, 0.07375 (2015-04-15 22:26:14) gbrt400 1212.csv')

d2 = diff_sets(s1,s2)
len(d2)

###3. 控制变量 gbrt n_estimators=400 和 500 的交集 （1212）

s1212_400 = pd.read_csv('data/output/gbrt/predict_set 1212 400.csv')
s1212_500 = pd.read_csv('data/output/gbrt/predict_set 1212 500.csv')

d3 = diff_sets(get_recommed_set(s1212_400), get_recommed_set(s1212_500))
len(d3)

###4. 控制变量 gbrt n_estimators=400 和 500 的交集 （~1212）

s_400 = pd.read_csv('data/output/gbrt/predict_set~1212 400.csv')
s_500 = pd.read_csv('data/output/gbrt/predict_set~1212 500.csv')

d4 = diff_sets(get_recommed_set(s_400), get_recommed_set(s_500))
len(d4)

##5. 10特征 （1212）

s_f10_500 = pd.read_csv('data/output/gbrt/f10_predict_set~1212 600 5.csv')
s_f10_600 = pd.read_csv('data/output/gbrt/f10_predict_set 1212 600 5.csv')
d5 = diff_sets(get_recommed_set(s_f10_500), get_recommed_set(s_f10_600))
len(d5)
